llm-otel-kit/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-otel-kit"
version = "0.1.0"
description = "Drop-in OpenTelemetry GenAI observability for any LLM backend — local or cloud."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Vishruth Harithsa" }]
keywords = ["opentelemetry", "llm", "observability", "genai", "dynatrace", "ollama", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "httpx>=0.27",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-http>=1.25",
    "traceloop-sdk>=0.59",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.25"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.urls]
Homepage = "https://github.com/theharithsa/Local-LLM-Application-with-OpenLLMetry"
Issues = "https://github.com/theharithsa/Local-LLM-Application-with-OpenLLMetry/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_otel_kit"]
